+++ GameServer/packets/Server/PacketLib1124.cs
@@ -368,7 +368,7 @@ namespace DOL.GS.PacketHandler
 				pak.WriteFloatLowEndian(playerToCreate.X);
 				pak.WriteFloatLowEndian(playerToCreate.Y);
 				pak.WriteFloatLowEndian(playerToCreate.Z);
-				pak.WriteShort((ushort)playerToCreate.Client.SessionID);
+				pak.WriteShort(playerToCreate.Client.SessionID);
 				pak.WriteShort((ushort)playerToCreate.ObjectID);
 				pak.WriteShort(playerToCreate.Heading);
 				pak.WriteShort(playerToCreate.Model);
